chore(pointcloud): factorize potree/potree2 nodes
